requirements.txt
streamlit
requests
pandas
python-dotenv
groq
tabulate